pyproject.toml
[tool.poetry]
name = "canvas"
version = "0.1.0"
description = "the motion canvas api server based on fastapi"
authors = ["Heewon Jeong <jhw07191@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
fastapi = "0.115.12"
uvicorn = "0.34.2"
azure-storage-blob = "12.25.1"
openai = "1.76.0"
annotated-types = "0.7.0"
anyio = "4.9.0"
azure-core = "1.33.0"
certifi = "2025.1.31"
cffi = "1.17.1"
charset-normalizer = "3.4.1"
click = "8.1.8"
cryptography = "44.0.2"
distro = "1.9.0"
h11 = "0.16.0"
httpcore = "1.0.9"
httpx = "0.28.1"
idna = "3.10"
isodate = "0.7.2"
pillow = "11.2.1"
pycparser = "2.22"
pydantic = "2.11.3"
pydantic-core = "2.33.1"
python-dotenv = "1.1.0"
requests = "2.32.3"
six = "1.17.0"
sniffio = "1.3.1"
starlette = "0.46.2"
tqdm = "4.67.1"
typing-inspection = "0.4.0"
typing-extensions = "4.13.2"
urllib3 = "2.4.0"
gunicorn = "^23.0.0"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
